fix parse errors display, refactor cli code
